+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- `erc_warnings` pre-flight option to consider ERC warnings as errors.
 
 ## [0.10.1] - 2021-02-22
 ### Added